+++ src/GoodByeContext.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const GoodByeContext = React.createContext();
+
+export default GoodByeContext;
